Removal of Mains
I removes main(string[args]) from both files "FilesAndClipboard.java"
and "MosueClick.java" because they are not mean to be run independently,
rather they are meant to be imported into other classes. I lefts the
mains as functions to be called later for troubleshooting. They still
exists as good ways of testing to see if the classes' code works.